requirements.txt
google-generativeai==0.3.2
configparser==6.0.1
